fixed mismatching integral types

+++ tests/std/tests/P0896R4_views_join/test.cpp
@@ -562,10 +562,10 @@ int main() {
     }
 
     { // Joining a non-const view without const qualified begin and end methods
-        vector<vector<size_t>> nested_vectors = {{0}, {1, 2, 3}, {99}, {4, 5, 6, 7}, {}, {8, 9, 10}};
-        auto RemoveSmallVectors       = [](const vector<size_t>& inner_vector) { return inner_vector.size() > 2; };
-        auto filted_and_joined        = nested_vectors | views::filter(RemoveSmallVectors) | views::join;
-        static constexpr int result[] = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10};
+        vector<vector<int>> nested_vectors = {{0}, {1, 2, 3}, {99}, {4, 5, 6, 7}, {}, {8, 9, 10}};
+        auto RemoveSmallVectors            = [](const vector<int>& inner_vector) { return inner_vector.size() > 2; };
+        auto filted_and_joined             = nested_vectors | views::filter(RemoveSmallVectors) | views::join;
+        static constexpr int result[]      = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10};
         assert(ranges::equal(filted_and_joined, result));
     }
